requirements.txt
streamlit>=1.30.0
openai>=1.10.0
jinja2>=3.1.0
python-dotenv>=1.0.0
python-docx>=1.1.0
chromadb>=0.4.0
